sumerian_mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sumerian_mcp"
version = "0.1.0"
authors = [
    {name = "Devramyun", email = "devramyun@gmail.com"},
]
description = "수메르어를 활용한 고급 암호화 시스템의 MCP 서버"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=0.1.0",
    "pycryptodome>=3.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
sumerian-mcp = "sumerian_mcp.cli:main"

[tool.setuptools]
packages = ["sumerian_mcp"] 